merges the aws,arlas and resto into one qssearch plugin
This plugin is a generalisation of search interfaces that perform search
through a query string.
This commit also changes the EOProduct API so that it takes the name of
the provider that provided it, the metadata, and the search parameters
to be built.
Along the way, all the search parameters supported by OpenSearch
extension for Earth Observation have been added in the default metadata
mapping.
Updates the providers making use of query string-like plugin to take
into account the new model.