changes the place where download url is resolved
It is now at the level of the download method of the EOProduct, and the
semantics has changed: we resolve the url with all the configuration
parameters of the download plugin 'as is', whithout any processing. This
particularly means that for HTTPDownload plugin, the config param
'base_uri' should be provided without a trailing '/' when it is present
in the EOProduct.remote_location in the form: '%(base_uri)s/something').