makes returning pagination info optional
a new kwarg is added in search method of eodag core api:
with_pagination_info and is False by default. This enables the user to
explicitly request for pagination information in a script while letting
the core api behave as before, not breaking existing code. The option is
set to True in the CLI, and is intended to be later extended