fixes bug with products having invalid geometries
Before now, a product with an invalid geometry was kept in search
results, making them not json-serializable (because of the way eodag
handles shapely's TopologicalError: the attr search_intersection is
given a value of None). This commit fixes this problem. This means even
products with invalid geometries are returned as search results