updates eodag search api for pagination
Simplifies the way eodag search for product types on the providers: the
partial_support mechanism is removed, simplifying the way pagination
works.
The search interface is modified to return a 2-tuple, the first item
being the result and the second the total number of items satisfying the
request.
The EOProduct properties now excludes all metadata that were either not
mapped or not available (mapped in the provider metadata_mapping but not
present in the provider response). This lowers the size of the number of
elements needed to be transferred as response to http requests for the
embedded http server.
Two new cli args are added: --page and --items to precise which page is
to be requested on the provider (default 1) and how many results to
retrieve (default 20).
The caching mechanism is removed on the embedded http server